[ui] Move remaining uses of “Modal” to “Dialog”
We’ve been a little inconsistent about this terminology (meaning I’ve been calling things Modal and everyone else is calling them Dialogs :-). This PR folds the remaining uses of Modal into Dialog.
