+++ js_modules/dagster-ui/packages/ui-core/src/assets/asset-checks/__stories__/AssetChecks.stories.tsx
@@ -12,7 +12,7 @@ import {
 } from '../../../graphql/types';
 import {AssetFeatureProvider} from '../../AssetFeatureContext';
 import {buildQueryMock} from '../../AutoMaterializePolicyPage/__fixtures__/AutoMaterializePolicyPage.fixtures';
-import {ASSET_CHECK_DETAILS_QUERY} from '../AssetCheckDetailModal';
+import {ASSET_CHECK_DETAILS_QUERY} from '../AssetCheckDetailDialog';
 import {AssetChecks} from '../AssetChecks';
 import {ASSET_CHECKS_QUERY} from '../AssetChecksQuery';
 import {
@@ -21,7 +21,7 @@ import {
   testAssetKey,
   testLatestMaterializationTimeStamp,
 } from '../__fixtures__/AssetChecks.fixtures';
-import {AssetCheckDetailsQueryVariables} from '../types/AssetCheckDetailModal.types';
+import {AssetCheckDetailsQueryVariables} from '../types/AssetCheckDetailDialog.types';
 import {AssetChecksQueryVariables} from '../types/AssetChecksQuery.types';
 
 // eslint-disable-next-line import/no-default-export
